+++ test/cases/chunkFilename-fullhash/expected/webpack-5/main.js
@@ -73,7 +73,7 @@ __webpack_require__.r(__webpack_exports__);
 /******/ 	
 /******/ 	/* webpack/runtime/getFullHash */
 /******/ 	(() => {
-/******/ 		__webpack_require__.h = () => ("04f5273a6b9819ed9e63")
+/******/ 		__webpack_require__.h = () => ("32b536df514e34b610fa")
 /******/ 	})();
 /******/ 	
 /******/ 	/* webpack/runtime/global */
